🏺 Added junction table artifact_sub_stat (#34)
