+++ backend/app/database/tables/__init__.py
@@ -1,17 +1,6 @@
 """Database Genshin Impact Characters Analyzer Models
 
-A package with descriptions of database entity tables.
-
-It describes the tables of the database used and the interfaces for interacting with them.
+A package with descriptions of database tables.
 """
 
-from .artifact import Artifact
 from .base import Base
-from .character import Character
-from .element import Element
-from .region import Region
-from .set import Set
-from .stat import Stat
-from .user import User
-from .user_character import UserCharacter
-from .weapon import Weapon
