+++ backend/app/database/tables/junctions/__init__.py
@@ -1,0 +1,9 @@
+"""Database Genshin Impact Characters Analyzer Junctions Models
+
+A package with descriptions of database junction tables.
+
+It describes the junction tables of the database used and the interfaces for interacting with them.
+"""
+
+from .artifact_sub_stat import ArtifactSubStat
+from .user_character import UserCharacter
